requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
ciso8601==2.1.3
cryptography==3.3.2
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.0.8
django-js-asset==1.2.2
djangorestframework==3.11.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytz==2021.1
PyYAML==5.4.1
requests==2.9.1
whitenoise==2.0.6
